+++ src/onegov.notice/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Allows to sort by group and user names.
+  [msom]
+
 0.1.2 (2017-08-30)
 ~~~~~~~~~~~~~~~~~~~